simple user permissions added

+++ main.py
@@ -14,20 +14,22 @@ templates = Jinja2Templates(directory='templates')
 @app.on_event('startup')
 async def run_on_startup():
     init_db()
-    pass
 
 
 @app.get('/', response_class=HTMLResponse)
 async def get_root(request: Request):
-    return templates.TemplateResponse('index.html', {'request': request, 'first': 'Hello', 'second': 'world!'})
+    return templates.TemplateResponse('index.html',
+                                      {'request': request, 'first': 'Hello', 'second': 'world!'})
 
 
 @app.post('/users/')
 async def create_user(user: UserModel):
     try:
         hashed_password = get_password_hash(user.hashed_password)
-        sql = "INSERT INTO users (username, first_name, last_name, birthday, gender, disabled, hashed_password) VALUES (?, ?, ?, ?, ?, ?, ?)"
-        values = (user.username, user.first_name, user.last_name, user.birthday, user.gender, user.disabled, hashed_password)
+        sql = ("INSERT INTO users (username, first_name, last_name, birthday, gender, disabled, hashed_password, "
+               "is_superuser) VALUES (?, ?, ?, ?, ?, ?, ?, ?)")
+        values = (user.username, user.first_name, user.last_name, user.birthday, user.gender, user.disabled,
+                  hashed_password, 0)
         db_worker('ins', sql, values)
         return user
 
@@ -36,35 +38,61 @@ async def create_user(user: UserModel):
 
 
 @app.get('/users/{user_id}')
-async def get_user(user_id: int):
+async def get_user(user_id: int, current_user: Annotated[UserModel, Depends(get_current_active_user)]):
     try:
-        sql = "SELECT username, first_name, last_name, birthday, gender FROM users WHERE id = ?"
-        values = (user_id, )
-        r = db_worker('fo', sql, values)
-        return {'username': r[0], 'first_name': r[1], 'last_name': r[2], 'birthday': r[3], 'gender': r[4]}
+        if user_id == current_user['id'] or current_user['is_superuser'] == 1:
+            sql = "SELECT username, first_name, last_name, birthday, gender FROM users WHERE id = ?"
+            values = (user_id, )
+            r = db_worker('fo', sql, values)
+            return {'username': r[0], 'first_name': r[1], 'last_name': r[2], 'birthday': r[3], 'gender': r[4]}
+        else:
+            return HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail='Resource is available but not accessible because of permissions',
+                headers={})
 
     except Exception as e:
-        return {'error': 'Not found'}
+        return {'error': e}
 
 
 @app.delete('/users/{user_id}')
-async def delete_user(user_id: int):
+async def delete_user(user_id: int, current_user: Annotated[UserModel, Depends(get_current_active_user)]):
     try:
-        sql = "DELETE FROM users WHERE id = ?"
-        values = (user_id, )
-        db_worker('del', sql, values)
+        if user_id == current_user['id'] or current_user['is_superuser'] == 1:
+            sql = "DELETE FROM users WHERE id = ?"
+            values = (user_id, )
+            db_worker('del', sql, values)
+        else:
+            return HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail='Resource is available but not accessible because of permissions',
+                headers={})
 
     except Exception as e:
         return {'error': e}
 
 
 @app.put('/users/{user_id}')
-async def update_user(user_id: int, user: UserModel):
+async def update_user(user_id: int, user: UserModel,
+                      current_user: Annotated[UserModel, Depends(get_current_active_user)]):
+    print(f'{user_id=}'
+          f'{user=}'
+          f'{current_user=}')
     try:
-        values = (user.username, user.first_name, user.last_name, user.birthday, user.gender, user_id)
-        sql = "UPDATE users SET username = ?, first_name=?, last_name=?, birthday=?, gender=? WHERE id=?"
-        db_worker('upd', sql, values)
-        return user
+        if user_id == current_user['id'] or current_user['is_superuser'] == 1:
+            values = (user.username, user.first_name, user.last_name, user.birthday, user.gender, user.is_superuser,
+                      user_id)
+            sql = ("UPDATE users SET username = ?, first_name=?, last_name=?, birthday=?, gender=?, is_superuser=?"
+                   "WHERE id=?")
+            db_worker('upd', sql, values)
+            return user
+
+        else:
+            return HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail='Resource is available but not accessible because of permissions',
+                headers={})
+
     except Exception as e:
         return {'error': e}
 
@@ -94,4 +122,5 @@ async def read_users_me(current_user: Annotated[UserModel, Depends(get_current_a
 
 @app.get("/users/me/details")
 async def read_own_items(current_user: Annotated[UserModel, Depends(get_current_active_user)]):
-    return [{'username': current_user['username'], 'first name': current_user['first_name'], 'last name': current_user['last_name'],}]
+    return [{'username': current_user['username'], 'first name': current_user['first_name'],
+             'last name': current_user['last_name'],}]
